moving files around added directories
