+++ data_structures/linkedlist.py
